requirements.txt
ollama
regex
langchain-google-genai
google-genai
supabase
python-dotenv
langchain-text-splitters
fastapi
uvicorn
anthropic
langgraph
requests
PyMuPDF
Pillow
pyyaml
